Make the Parser the parent of all new Nodes.
After creating a node, a NodeFactory may instruct the parser to parse more of the template.
If that parsing stage throws an exception, the new Node does not have a parent yet, so will
not be deleted. By making the Parser the parent of the new Nodes, they will be deleted even if
an exception is thrown.

An example scenario where this matters is when a filter does not exist and an exception is thrown.

"{% if foo %}{% if bar %}{{ bat|does_not_exist }}{% endif %}{% endif %}"