Change SafeString to make the QString available through operator->().
Although it works on gcc, operator QVariant() is not preferred
over /* implicit */ QVariant(QString) on MSVC 2008, sun studio
and possibly others.

This derived-to-base behaviour is worked around by making SafeString
not a sub class of QString anymore. QString api is accessible through
operator->() and parts of it will be overridden later.

This includes a test file for testing the behaviour of different compilers
with different work arounds attempted.