Add a way to build Grantlee with Qt 5.
This feature is provided for temporary porting convenience. It allows
downstreams to proceed with the more-difficult parts of their port
to Qt 5 without imposing the minor breaking changes from the next
version of Grantlee. Porting to the next version of Grantlee is a
necessary part of such a port however because this option produces
a binary-incompatible version of Grantlee with SOVERSION 0, and
because it does not make use of the new Qt 5 features. This only
allows a 'dumb' port.