Allow the metatype system to work with Qt 5.
Disable infinite recursion while compiling by defining the
GRANTLEE_REGISTER_SEQUENTIAL_CONTAINER_IF and
GRANTLEE_REGISTER_ASSOCIATIVE_CONTAINER_KEY_IF macros to
empty in Qt 5. The task of those macros is not necessary
with Qt 5 built-in container handling.

Use the built-in container handling.

Register custom containers in tests with the Qt abstractions in Qt 5
instead of the grantlee abstractions, so that the built-in facilities
work for those.