Don't put template into invalid state if included template is not found
When using the CachingLoaderDecorator the template evaluation
suffered from a memory effect. When one template had a line
like {% include var_for_dynamic_template %} where each render context
had a different template path, things could simply stuck.
Once a previous render() had a value that didn't match a template
file then the cache marks the template which has the include tag
as on error. This made all next calls render() fails as the template
was on error, when not using CachingLoaderDecorator the template
is reevaluated every time so after a failed include the next
render() would still work because it would be completely parsed again.
In short a include line without syntax errors should't make your
template be in an invalid state even if the included file does not exist.