Template: Clear errors after successfully rendering
If a template is rendered with a particular Context, where something in
that Context results in an error, that error is stored on the template.

It should be possible to render the template again with a different
Context, this time resulting in no error.  However, as the previous
error is still stored on the Template object, the rendering still
reports the previous error, despite successful rendering.

Fix this by clearing the error state when the rendering is successful.

Adjust unit tests to check for errors when a template is created.  The
unit tests should always have done this, but were too lazy previously.

It is an unfortunate design that the error state is stored on the
Template object.  It would be better if a separate ErrorStorage object
were required in the render() interface.