Parser: Consider only the command name to determine block validity
Don't attempt to match the entire token contents to determine the
termination condition.  Refactor the block tag to perform error checking
after parsing, as is done by other tags like ifchanged.  The block tag
is the only built-in tag which expects particular content in its
end-tag.

This relaxation of the termination condition means that it is possible
to implement the elif tag as a companion to the if tag.  The elif tag
takes arguments, but the termination condition should be the elif
command name only.  The arguments to it are not pre-determined.

 https://code.djangoproject.com/ticket/3100