Lowering the needed CMAKE Version
The latest CMAKE Version available in Jessie is CMAKE 3.0.2 (see
https://packages.debian.org/de/jessie/cmake). This package is used to build KDE
Framework 5, wich is used to build Kleopatra
(https://www.kde.org/applications/utilities/kleopatra/). Currently, this is the
only dependency to build it that requires a package, that is not part the Debian
stable. This lowers the needed Version to 3.0 and removes all its dependencies.