for: Remove dead code
This code has been inactive since commit v0.1.0-rc1~371 (Remove
dictionary unpack feature. SiC., 2009-06-09).

The feature was removed because Django doesn't allow iterating over
associative containers directly, requiring the user to use ".items", as
is also required in the python language.  I left the dead code as a
reminder to myself to consider trying to upstream a change to that
semantic, but doing so probably does not make sense at this point, given
that Django is just following python.