Add filesizeformat to the the default string filters.
formatfilesize can be used like the counterpart from Django, but
has some more options. It features decimal units as well as binary
units. Decimal precision can directly be set in the filter. For
situations where the input value is not byte, but something like
kB or KiB, you can add a multipliert for that base value.

filesizeformat accepts three optional parameters:
1. the unit system: either 2 for binary unity like KiB or the default
   10 for decimal units like kB
2. the precision for the decimal output number, the default is 2
3. an multiplier for input values that are not bytes, but something
   like KiB - than simply set a multiplier of 1024

Some examples:
{{ my_file_size|filesizeformat }}
will simply work like the filter in django and expects a byte value
that will be shown ad decimal with a precision of 2, like 23,45 MB

{{ my_file_size|filesizeformat:"2" }}
will output in binary units like 1,24 KiB

{{ my_file_size|filesizeformat:"10,2,1000" }}
will output decimal units with a precision of 2 where the input value
is in KB.