Implement l10n_filesize and l10n_filesize_var
These 2 tags are like localized versions of the filesizeformat filter.
The filter only outputs values with a dot as decimal separator like used
in the english language. These two tags will convert the file size value
into a localized number string. The have the same options as the
filesizeformat filter has.

{% l10n_filesize value unitSystem precision multiplier %}

{% l10n_filesize_var value unitSystem precision multiplier varName %}

unitSystem can be either 2 for binary units or 10 (the default) for
decimal units

precision is the decimal precision after the separator

the multiplier can be used if the input value is not in pure bytes but
in something like KiB or KB. If you have your input value only available
in KiB for example, simply use a multiplier of 1024 to get the base byte
value. The default multiplier is 1.