Explicitly tests truthiness for float
Explicitly tests the truthiness for float values in addition to double
ones; related to this, switch the double variable from auto back to the
explicit type, to stress the difference (and make sure it is always
handled as double).

Very late followup of fbd54ee9a326cb1b1c438010882d02b1eaa08259.