filesizeformat implementation as well as localized filesize (#2)
* filesizeformat implementation as well as localized filesize

This implements the filesizeformat filter as it is available
in Django (plus some optional features). Additionally it
introduces two new tags (l10n_filesize and l10n_filesize_var)
for the i18n system to support file sizes with localized number
strings.

Originally from: https://github.com/steveire/grantlee/pull/36

* Use QLocale::formattedDataSize() if available (Qt 5.10+)

If build against Qt 5.10 or newer, internally QLocale::formattedDataSize()
will be used for values that fit into a qint64. The old implementation
is still there for older Qt versions and for values that do not fit into
a qint64.

The main difference bitween the internal implementation in utils.cpp as
calcFileSize() and QLocale::formattedDataSize() is that QLocale also
translates the unit name.